add precompile specific contract tests
